Remove use of const_cast and Eigen::Refs
Revert to using pointers where appropriate instead of Eigen::Refs which
are used incorrectly.

+++ example/ex01_essential_estimation.py
@@ -71,7 +71,8 @@ def normalize_to_ubyte(x):
 
 def step1_sift_detect(args):
     """Run SIFT key-point detection and descriptors on images."""
-    ims = [imread(image_filename, force_grayscale=True)
+    ims = [imread(image_filename, dtype='float32',
+            force_grayscale=True)
            for image_filename in args.images]
     with Timer('step1-computation'):
         siftkps = [sift_filter(im) for im in ims]
